SEDA4BPEL/zmq-eda/requirements.txt
PyYAML==3.11
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.18
celery==3.1.12
flower==0.7.0
kombu==3.0.21
pytz==2014.4
pyzmq==14.3.1
tornado==3.2.2
wsgiref==0.1.2
